+++ src/mixins/ids-hitbox-mixin/index.js
@@ -1,0 +1,1 @@
+export { default as IdsHitboxMixin } from './ids-hitbox-mixin';