+++ src/mixins/ids-popup-interactions-mixin/ids-popup-interactions-mixin.js
@@ -14,7 +14,7 @@ const POPUP_TRIGGER_TYPES = [
  * @param {any} superclass Accepts a superclass and creates a new subclass from it
  * @returns {any} The extended object
  */
-const IdsPopupInteractionsMixin = (superclass) => class extends IdsEventsMixin(superclass) {
+const IdsPopupInteractionsMixin = (superclass) => class extends superclass {
   constructor() {
     super();