+++ src/mixins/ids-tooltip-mixin/ids-tooltip-mixin.js
@@ -1,5 +1,4 @@
 import { attributes } from '../../core';
-import { IdsEventsMixin } from '../ids-events-mixin';
 import IdsTooltip from '../../components/ids-tooltip';
 
 /**
@@ -8,7 +7,7 @@ import IdsTooltip from '../../components/ids-tooltip';
  * @param {any} superclass Accepts a superclass and creates a new subclass from it
  * @returns {any} The extended object
  */
-const IdsTooltipMixin = (superclass) => class extends IdsEventsMixin(superclass) {
+const IdsTooltipMixin = (superclass) => class extends superclass {
   constructor() {
     super();
   }